+++ triplea/service/repository/state/__init__.py
@@ -1,4 +1,5 @@
 from triplea.service.repository.state.expand_details import expand_details
+from triplea.service.repository.state.get_full_text import get_full_text
 from triplea.service.repository.state.parsing_details import parsing_details
 
 # from triplea.service.repository.state.ner_title import ner_title
@@ -24,6 +25,7 @@ __all__ = [
     "expand_details",
     "parsing_details",
     "get_citation",
+    "get_full_text",
     "get_article_list_from_pubmed_all_store_to_arepo",
     "extract_triple_abstract_save",
     "affiliation_mining",